Move printing functions and stats functions (global stats) to a separate file
