+++ CHANGELOG.md
@@ -10,7 +10,8 @@ Next
   recvmsg with MSG_DONTWAIT instead (#449)
 - Improved compatibility with NetBSD (#452, thanks @auerswal)
 - Consistent variable names for command-line options (#453, thanks @gsnw-sebast)
-- Move the most function declarations from fping.c to fping.h (#454, thanks @gsnw-sebast)
+- Move printing functions, stats functions (global stats) and
+  function declarations to a separate file (#454, thanks @gsnw-sebast)
 - Avoid crash with option `-0` (#457, thanks @auerswal)
 
 fping 5.5 (2025-12-31)
