+++ src/stats.h
@@ -1,0 +1,10 @@
+#ifndef _STATS_H
+#define _STATS_H
+
+#include "fping.h"
+
+void print_per_system_stats(void);
+void print_per_system_stats_json(void);
+void print_global_stats(void);
+void print_global_stats_json(void);
+#endif
